Refactor nom debug_type et modification pour suivre celles au niveau de Coach.py de StrategyIA
